swiper 구현 : feat : MovingCardDocument 구현 https://github.com/Sejong-Balsamic/Malsami-FE/issues/479
