+++ src/components/common/Card.tsx
@@ -48,7 +48,7 @@ export default function Card({
 }: CardProps) {
   return (
     <article
-      className="flex h-[194px] w-[261px] flex-col justify-between rounded-[20px] p-[20px] shadow-md hover:shadow-lg"
+      className="flex h-[194px] w-[261px] flex-col justify-between rounded-[20px] bg-white p-[20px] shadow-md hover:shadow-lg"
       onClick={onClick}
     >
       <>
